add XMLDataset and make VOCDataset a child of it
