+++ mmdet/datasets/__init__.py
@@ -1,4 +1,5 @@
 from .custom import CustomDataset
+from .xml_style import XMLDataset
 from .coco import CocoDataset
 from .voc import VOCDataset
 from .loader import GroupSampler, DistributedGroupSampler, build_dataloader
@@ -7,7 +8,7 @@ from .concat_dataset import ConcatDataset
 from .repeat_dataset import RepeatDataset
 
 __all__ = [
-    'CustomDataset', 'CocoDataset', 'VOCDataset', 'GroupSampler',
+    'CustomDataset', 'XMLDataset', 'CocoDataset', 'VOCDataset', 'GroupSampler',
     'DistributedGroupSampler', 'build_dataloader', 'to_tensor', 'random_scale',
     'show_ann', 'get_dataset', 'ConcatDataset', 'RepeatDataset'
 ]
